requirements.txt
python-telegram-bot==21.7
APScheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.2
feedparser==6.0.11
pytest==8.3.3
